roll premission implement

+++ app/Http/Controllers/Admin/Admin/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UpdateRoleRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
@@ -15,6 +16,9 @@ class RoleController extends Controller
      */
     public function index()
     {
+        if(!Auth::user()->can('role list')){
+            abort(403);
+        }
         $roles = Role::latest()->paginate(10);
         // return  $roles;
         return view('admin.role.index',compact('roles'));
@@ -25,6 +29,10 @@ class RoleController extends Controller
      */
     public function create()
     {
+
+        if(!Auth::user()->can('role create')){
+            abort(403);
+        }
         $roles = Role::all();
         // return $roles;
         $permissions = Permission::all()->groupBy('module_name');
@@ -54,7 +62,9 @@ class RoleController extends Controller
      */
     public function show(string $id)
     {
-        //
+        if(!Auth::user()->can('role show')){
+            abort(403);
+        }
     }
 
     /**
@@ -62,7 +72,9 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
-
+        if(!Auth::user()->can('role update')){
+            abort(403);
+        }
         $role = Role::where('id', $id)->first();
         $roles = Role::all();
         $permissions = Permission::all()->groupBy('module_name');
@@ -77,6 +89,9 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, Role $role)
     {
+        if(!Auth::user()->can('role update')){
+            abort(403);
+        }
         $role->update(['name' => $request->validated('name')]);
         $role->syncPermissions( array_map('intval', $request->validated('permission_ids')));
         return redirect()->route('role.index');
@@ -87,6 +102,9 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        if(!Auth::user()->can('role delete')){
+            abort(403);
+        }
         $role->delete();
         return redirect()->route('role.index');
     }
